move nika files to nika mod
