+++ src/lib.rs
@@ -1,3 +1,2 @@
 pub mod keyboards;
-pub mod nika_client;
-pub mod nika_response;
+pub mod nika;
